Rename gotham to newyork, remove authorizable comments
