--- gotham-server/src/main.rs
@@ -1,14 +1,0 @@
-// Gotham-city
-//
-// Copyright 2018 by Kzen Networks (kzencorp.com)
-// Gotham city is free software: you can redistribute
-// it and/or modify it under the terms of the GNU General Public
-// License as published by the Free Software Foundation, either
-// version 3 of the License, or (at your option) any later version.
-//
-
-use server_lib::server;
-
-fn main() {
-    server::get_server().launch();
-}
