--- gotham-server/src/routes/mod.rs
@@ -1,13 +1,0 @@
-// Gotham-city
-//
-// Copyright 2018 by Kzen Networks (kzencorp.com)
-// Gotham city is free software: you can redistribute
-// it and/or modify it under the terms of the GNU General Public
-// License as published by the Free Software Foundation, either
-// version 3 of the License, or (at your option) any later version.
-//
-
-pub mod ecdsa;
-pub mod schnorr;
-pub mod eddsa;
-pub mod ping;
