+++ integration-tests/Cargo.toml
@@ -12,11 +12,11 @@ time-test = "0.2.1"
 log = "0.4"
 time = "*"
 
-[dependencies.gotham-server]
-path = "../gotham-server"
+[dependencies.newyork-server]
+path = "../newyork-server"
 
-[dependencies.gotham-client]
-path = "../gotham-client"
+[dependencies.newyork-client]
+path = "../newyork-client"
 
 [patch.crates-io]
 rust-gmp = { version = "0.5.0", features = ["serde_support"], git = "https://github.com/KZen-networks/rust-gmp" }
