+++ integration-tests/README.md
@@ -1,6 +1,6 @@
 # Integration tests
 
-Runs the Gotham-server and calls it using Gotham-client's API.<br>
+Runs the Newyork-server and calls it using Newyork-client's API.<br>
 Run it simply with:
 ```sh
 $ cargo test
