+++ launch-server.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-cd ./gotham-server && cargo run --release
+cd ./newyork-server && cargo run --release
