+++ newyork-server/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gotham-server"
+name = "newyork-server"
 authors = [
     "gbenattar <g.benattar@gmail.com>",
     "Oded Leiba <odedleiba5@gmail.com"
