+++ newyork-server/README.md
@@ -1,14 +1,14 @@
-# Gotham Server
-![Gotham Server](../misc/server-icon.png)
+# Newyork Server
+![Newyork Server](../misc/server-icon.png)
 
 ## Introduction
-Gotham server is a RESTful web service exposing APIs for two party ECDSA key generation and signing.
+Newyork server is a RESTful web service exposing APIs for two party ECDSA key generation and signing.
 
 ## Installation
 ### Launching the server
 ```bash
-git clone https://github.com/KZen-networks/gotham-city.git
-cd gotham-city/gotham-server
+git clone https://github.com/KZen-networks/newyork-city.git
+cd newyork-city/newyork-server
 cargo run --release
 ```
