+++ newyork-server/src/auth/jwt.rs
@@ -2,14 +2,7 @@ use std::collections::HashSet;
 
 use jsonwebtoken::DecodingKey;
 
-// Gotham-city
-//
-// Copyright 2018 by Kzen Networks (kzencorp.com)
-// Gotham city is free software: you can redistribute
-// it and/or modify it under the terms of the GNU General Public
-// License as published by the Free Software Foundation, either
-// version 3 of the License, or (at your option) any later version.
-//
+ 
 use super::super::jwt::{decode, decode_header, Algorithm, Header, Validation};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
