+++ newyork-server/src/lib.rs
@@ -1,11 +1,4 @@
-// Gotham-city
-//
-// Copyright 2018 by Kzen Networks (kzencorp.com)
-// Gotham city is free software: you can redistribute
-// it and/or modify it under the terms of the GNU General Public
-// License as published by the Free Software Foundation, either
-// version 3 of the License, or (at your option) any later version.
-//
+ 
 #![recursion_limit = "128"]
 #![feature(proc_macro_hygiene)]
 #![feature(decl_macro)]
