+++ newyork-server/src/main.rs
@@ -1,0 +1,7 @@
+ 
+
+use server_lib::server;
+
+fn main() {
+    server::get_server().launch();
+}
