+++ newyork-server/src/routes/eddsa.rs
