+++ newyork-server/src/routes/mod.rs
@@ -1,0 +1,6 @@
+ 
+
+pub mod ecdsa;
+pub mod schnorr;
+pub mod eddsa;
+pub mod ping;
