+++ newyork-server/src/routes/ping.rs
