+++ newyork-server/src/routes/schnorr.rs
