+++ newyork-utilities/server/buildspec.yml
@@ -14,7 +14,7 @@ phases:
     commands:
     - echo Build started on `date`
     - echo Building the Docker image...
-    - docker build -t $REPOSITORY_URI:latest . -f ./gotham-utilities/server/docker-build-img/Dockerfile
+    - docker build -t $REPOSITORY_URI:latest . -f ./newyork-utilities/server/docker-build-img/Dockerfile
     - docker tag $REPOSITORY_URI:latest $REPOSITORY_URI:$IMAGE_TAG
   post_build:
     commands:
